Removed unused .css files.  Added the new SinglePropertyPage component.  Added the route for the single property page.
