--- src/components/HomePage/HomePage.css
@@ -1,2 +1,0 @@
-.home-page {
-}
