--- src/components/PropertyList/PropertyList.css
@@ -1,2 +1,0 @@
-.property-list {
-}
