+++ src/components/PropertyListItem/PropertyListItem.js
@@ -16,7 +16,6 @@ limitations under the License.
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import { Link } from "react-router-dom";
-import './PropertyListItem.css';
 import Button from '@material-ui/core/Button';
 import Card from '@material-ui/core/Card';
 import CardActions from '@material-ui/core/CardActions';
