--- src/components/SingleProperty/SingleProperty.css
@@ -1,2 +1,0 @@
-.property-list-item {
-}
